requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic[email]==2.4.2
transformers==4.35.0
torch==2.1.0
firebase-admin==6.2.0
requests==2.31.0
python-dotenv==1.0.0 